requirements.txt
# AI and ML
llama-cpp-python>=0.2.34
torch>=2.1.0
transformers>=4.36.0
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.2
reportlab>=4.0.0
Pillow>=10.0.0

# Web scraping and automation
selenium>=4.16.0
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Security tools integration
python-nmap>=0.7.1
scapy>=2.5.0
paramiko>=3.3.0
pymetasploit3>=1.0.4

# Data processing
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0

# Report generation
jinja2>=3.1.0
markdown>=3.5.0
weasyprint>=60.0
reportlab>=4.0.0

# Utilities
colorama>=0.4.6
tqdm>=4.66.0
click>=8.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
psutil>=5.9.0

# Network and security
netaddr>=0.9.0
dnspython>=2.4.0
cryptography>=41.0.0
